backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
mysql-connector-python
PyJWT
reportlab
python-dotenv
Pillow
matplotlib
opencv-python
pyzbar
numpy